Replace old spline implementation with new one
