+++ src/main/java/org/frcteam2910/common/math/spline/SplineHelper.java
@@ -1,4 +1,4 @@
-package org.frcteam2910.common.math.spline2;
+package org.frcteam2910.common.math.spline;
 
 import org.ejml.simple.SimpleMatrix;
