+++ src/test/java/org/frcteam2910/common/math/spline/CubicBezierSplineTest.java
@@ -1,4 +1,4 @@
-package org.frcteam2910.common.math.spline2;
+package org.frcteam2910.common.math.spline;
 
 import org.frcteam2910.common.math.Rotation2;
 import org.frcteam2910.common.math.Vector2;
@@ -8,8 +8,29 @@ import static org.junit.Assert.assertEquals;
 
 public class CubicBezierSplineTest {
     @Test
+    public void verifyStartEndPosition() {
+        CubicBezierSpline spline = new CubicBezierSpline(
+                new Vector2(0, 0),
+                new Vector2(5, 0),
+                new Vector2(45, 50),
+                new Vector2(50, 50)
+        );
+        assertEquals("Starting position is not correct", new Vector2(0, 0), spline.getPoint(0.0));
+        assertEquals("Ending position is not correct", new Vector2(50, 50), spline.getPoint(1.0));
+
+        spline = new CubicBezierSpline(
+                new Vector2(-20, 35),
+                new Vector2(10, 35),
+                new Vector2(40, -20),
+                new Vector2(50, -20)
+        );
+        assertEquals("Starting position is not correct", new Vector2(-20, 35), spline.getPoint(0.0));
+        assertEquals("Ending position is not correct", new Vector2(50, -20), spline.getPoint(1.0));
+    }
+
+    @Test
     public void verifyStartEndHeading() {
-        Spline spline = new CubicBezierSpline(
+        CubicBezierSpline spline = new CubicBezierSpline(
                 new Vector2(0, 0),
                 new Vector2(5, 0),
                 new Vector2(45, 50),
